Collapsed packet client handler code back into packet definition
